requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
blinker==1.7.0
boto3==1.34.62
botocore==1.34.62
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cohere==4.47
dataclasses-json==0.6.4
deprecation==2.1.0
distro==1.9.0
exceptiongroup==1.2.0
faiss-cpu==1.7.4
fastavro==1.9.4
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.42
gotrue==2.1.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.3
httpx==0.25.2
idna==3.6
importlib-metadata==6.11.0
Jinja2==3.1.3
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
langchain==0.1.7
langchain-community==0.0.20
langchain-core==0.1.23
langchain-openai==0.0.6
langsmith==0.0.87
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.20.2
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.12.0
opencv-python==4.9.0.80
packaging==23.2
pandas==2.2.0
pillow==10.2.0
postgrest==0.15.0
protobuf==4.25.3
pyarrow==15.0.0
pydantic==2.6.1
pydantic_core==2.16.2
pydeck==0.8.1b0
Pygments==2.17.2
PyMuPDF==1.23.26
PyMuPDFb==1.23.22
python-dateutil==2.8.2
python-pptx==0.6.23
pytz==2024.1
PyYAML==6.0.1
rank-bm25==0.2.2
realtime==1.0.2
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rich==13.7.0
rpds-py==0.18.0
s3transfer==0.10.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.27
storage3==0.7.0
streamlit==1.31.1
StrEnum==0.4.15
supabase==2.3.5
supafunc==0.3.3
tenacity==8.2.3
tiktoken==0.6.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
urllib3==1.26.18
validators==0.22.0
watchdog==4.0.0
websockets==11.0.3
XlsxWriter==3.2.0
yarl==1.9.4
zipp==3.17.0
